refactor(neuron-wallet): move decorators to the top-level folder
