+++ packages/neuron-wallet/src/decorators/errors.ts
