+++ packages/neuron-wallet/src/decorators/index.ts
@@ -1,0 +1,7 @@
+import errorsDecorators from './errors'
+
+export const { CatchControllerError } = errorsDecorators
+
+export default {
+  ...errorsDecorators,
+}
